refactor webcam util
